Making RTF Annotation labels based on annotation template

Making RTF Annotation labels based on annotation template

annotation_label(dat = NULL, spellcheck = TRUE, outfile = "Annotation_Labels.rtf")

Arguments

dat
data frame of Annotation Template.

spellcheck
Logical, whether the spell checking for species should be porformed.

outfile
the names of the output rtf file.

Details

This function reads csv or data frame "annoation template", and generates RTF file containing annotation labels.

PROJECT: project related with this annotation. eg. Examined for the Revision for the Fern Flora of Hong Kong

TYPE_STATUS: which kind of type specimen: eg. ISOTYPE

TYPE_REF: what reference the identification was based upon. DC. Monogr. Phanerog. 6: 409. 1889.

FAMILY: Family, eg. Magnoliaceae

GENUS:Genus, eg. Ophioglossum

SPECIES:Species, eg. petiolatum

AUTHOR_OF_SPECIES: Author, eg. Hook.

INFRASPECIFIC_RANK: Infra-specific Rank, eg. var., f. or subsp.

INFRASPECIFIC_EPITHET: Epithet for the infra specific rank.

AUTHOR_OF_INFRASPECIFIC_RANK: Author for the infra specific rank.

ABBREVIATION: Role of the person annotating this herbarium specimen, eg. "Det.", "Verified by".

IDENTIFIED_BY:Who identified the species, eg. Jin-Long Zhang

INSTITUTION:The institution where the identification was made, eg. KFBG

DATE_IDENTIFIED: should follow the format: 2014-12-19 for "19 December 2014".

DET_NOTE:The specimen is extremely close to Ranunculus japonicus

SPECIMEN_NUMBER: Global Unique Identifier, see the help file for herbarium_label for more information.

COLLECTOR: Collector of this specimen

COLLECTOR_NUMBER: collector's number of this specimen.

Author

Jinlong Zhang jinlongzhang01@gmail.com

Examples

    path <- system.file("extdata", "ANNOTATION_TEMPLATE.xlsx", 
                       package = "herblabel")
    library(openxlsx)
    dat <- read.xlsx(path)                    
    annotation_label(dat)

#> Warning: "GENUS" not provided for row:  6, 15, 16, 17, 18, 19, 20, 21, 23

#> Warning: "IDENTIFIED_BY" not provided for row:  14, 22

#> [1] "Warning: Date format incorrect, using original string"
#> [1] "Warning: Date format incorrect, using original string"
#> [1] "Warning: Date format incorrect, using original string"
#> Annotation labels have been saved to:
#> C:/Jinlong/github packages/herblabel/docs/reference/Annotation_Labels.rtf