Creating RTF herbarium labels based on Darwin Core herbarium records template

Creating RTF herbarium labels based on Darwin Core herbarium records template.

herbarium_label(dat = NULL, spellcheck = TRUE, theme = c("KFBG", "PE", "KUN", "HU"), 
                outfile = "herblabel.rtf", font = c("Roman","Arial"), font_size = 1)

Arguments

dat
dataframe of Darwin Core Template.

spellcheck
Logical, indicating whether the check of spelling should be performed.

theme
a character string speficying the theme of herbarium labels

outfile
Character string, name of the rtf file to be generated.

font
Font for all the labels to be generated,either "Roman" for proportionally spaced serif fonts or "Arial", proportionally spaced sans serif fonts respectively.

font_size
Relative Font Size, by default it is 1.

Details

This function reads Darwin Core herbarium records template in CSV, and generates a RTF file. It will check the completeness for each entry: Labels will not generated if some critical fields is missing, eg. Collector, Collector's Number. Therefore, a specimen without collector's number, the collector's number should be marked as "s.n.".

Please follow the APGIII families for all the species, if mismatch is found, there will be a warning in the RTF label.

The validity of the species will be checked based on the Plant List Website http://www.theplantlist.org/ or Flora of China http://frps.eflora.cn/ . So make sure the species is spelled correctly, and the Abbreviation of Authors is correctly spelled.

The function will also identify and highlight the latin words in the Remarks field.

Details for the columns of the template. please look at the file "DARWIN_CORE_HERBARIUM_RECORDS_TEMPLATE.xlsx" in the extdata folder:

1. GLOBAL_UNIQUE_IDENTIFIER: Institution code plus the number, eg. KFBG000102

2. INSTITUTION_CODE: Code for the herbarium, eg. KFBG

3. COLLECTION_CODE: the collection code for the herbarium specimen, eg. the barcode number

4. BASIS_OF_RECORD: type of the specimen, eg. herbarium specimen

5. PREPARATIONS:fruiting, flowering or vegetative, eg. fruiting

6. HERBARIUM:Name of the herbarium, eg. Kadoorie Farm and Botanic Garden Herbarium (KFBG).

7. TITLE: Title of the collection, eg. Plants of Hong Kong

8. COLLECTOR:Full Name of the collector.

9. ADDITIONAL_COLLECTOR:People assisted the collection.

10. COLLECTOR_NUMBER:Collector's number

11. DATE_COLLECTED:eg. Must must follow the YYYY-MM-DD format, eg. 2013-4-11, representing 11 April 2013.

12. LOCAL_NAME:Local name for this plant.

13. FAMILY:APGIII Family

14. GENUS:Genus

15. SPECIES:specific epithetic

16. AUTHOR_OF_SPECIES:author

17. INFRASPECIFIC_RANK: Infra-specific rank, one of "var.", "f.", or "subsp."

18. INFRASPECIFIC_EPITHET:infraspecific epithet

19. AUTHOR_OF_INFRASPECIFIC_RANK:author of the infraspecific epithet

20. COUNTRY:country of the specimen collected, must be provided.

21. STATE_PROVINCE:state or province, must be provided.

22. COUNTY: county

23. LOCALITY: precise locality

24. LOCALITY_ORIGINAL: locality in local language.

25. IMAGE_URL: url for the image of this specimen.

26. RELATED_INFORMATION: the reference of the name

27. LAT_DEGREE:latitude, degrees, in WGS1984 geocoordinate system.

28. LAT_MINUTE:latitude, minutes, in WGS1984 geocoordinate system.

29. LAT_SECOND:latitude, seconds, in WGS1984 geocoordinate system.

30. LAT_FLAG:N for northern hemisphere, S for southern hemisphere

31. LON_DEGREE:longitude, degrees, in WGS1984 geocoordinate system.

32. LON_MINUTE:longitude, minutes, in WGS1984 geocoordinate system.

33. LON_SECOND:longitude, seconds, in WGS1984 geocoordinate system.

34. LON_FLAG:E for eastern hemisphere, W for western hemisphere

35. ELEVATION:elevation in meters, please only provide the number.

36. ATTRIBUTES:description of the species. eg. Tree, 8m tall. Bark brownish. Young branchlet covered with dense hair. Leaf Underside greenish grey. Anthers yellow, petals red. Fruit 8-12mm, globose, yellow.

37. REMARKS:the associated species, and the description of habitats. Eg. Understory, with Lasianthus fordii, Ardisia quinquegona etc.

38. CABINET:location of the specimen

39. DISPOSITION:eg. in herbarium

40. GEOREFERENCE_SOURCES:GPS or Google Earth

41. GEOREFERENCE_VERIFICATION_STATUS:whether the georeferecing data has been veryfied or not. eg. YES

42. GEOREFERENCE_REMARKS:accuracy for the georeference

43. PROJECT:Project the specimen related with

44. IDENTIFIED_BY: who did the identification

45. DATE_IDENTIFIED: date of identification, eg. 2015-10-18

46. TYPE_STATUS:if the specimen is a type specimen or not. if yes, please specify which kind of type specimen it belong to.

47. PROCESSED_BY: data entered by

48. DATE_LASTMODIFIED: date of the modification, eg. 2015-10-18

References

The Plabel program developed by Kent D. Perkins at the University of Florida Herbarium http://www.flmnh.ufl.edu/natsci/herbarium/pl/ .

Author

Jinlong Zhang jinlongzhang01@gmail.com

Examples

    path <- system.file("extdata", "DARWIN_CORE_HERBARIUM_RECORDS.xlsx", 
                        package = "herblabel")
    library(openxlsx)
    dat <- read.xlsx(path)
    
    herbarium_label(dat, theme = "KFBG", outfile = "herbarium_labels_KFBG.rtf")

#> Warning: "FAMILY" not provided for row:  8

#> [1] "17 herbarium specimen labels to create:"
#> [1] "Making label for row:  1"
#> [1] "Making label for row:  5"
#> [1] "Making label for row:  10"
#> [1] "Making label for row:  15"
#> Herbarium Labels have been saved to:
#> C:/Jinlong/github packages/herblabel/docs/reference/herbarium_labels_KFBG.rtf

    herbarium_label(dat, theme = "PE",   outfile = "herbarium_labels_PE.rtf")

#> Warning: "FAMILY" not provided for row:  8

#> [1] "17 herbarium specimen labels to create:"
#> [1] "Making label for row:  1"
#> [1] "Making label for row:  5"
#> [1] "Making label for row:  10"
#> [1] "Making label for row:  15"
#> Herbarium Labels have been saved to:
#> C:/Jinlong/github packages/herblabel/docs/reference/herbarium_labels_PE.rtf

    herbarium_label(dat, theme = "KUN",  outfile = "herbarium_labels_KUN.rtf")

#> Warning: "FAMILY" not provided for row:  8

#> [1] "17 herbarium specimen labels to create:"
#> [1] "Making label for row:  1"
#> [1] "Making label for row:  5"
#> [1] "Making label for row:  10"
#> [1] "Making label for row:  15"
#> Herbarium Labels have been saved to:
#> C:/Jinlong/github packages/herblabel/docs/reference/herbarium_labels_KUN.rtf

    herbarium_label(dat, theme = "HU",  outfile = "herbarium_labels_HU.rtf")

#> Warning: "FAMILY" not provided for row:  8

#> [1] "17 herbarium specimen labels to create:"
#> [1] "Making label for row:  1"
#> [1] "Making label for row:  5"
#> [1] "Making label for row:  10"
#> [1] "Making label for row:  15"
#> Herbarium Labels have been saved to:
#> C:/Jinlong/github packages/herblabel/docs/reference/herbarium_labels_HU.rtf

    herbarium_label(dat, spellcheck = FALSE, 
                    outfile = "herbarium_labels_no_checking.rtf")

#> Warning: "FAMILY" not provided for row:  8

#> [1] "17 herbarium specimen labels to create:"
#> [1] "Making label for row:  1"
#> [1] "Making label for row:  5"
#> [1] "Making label for row:  10"
#> [1] "Making label for row:  15"
#> Herbarium Labels have been saved to:
#> C:/Jinlong/github packages/herblabel/docs/reference/herbarium_labels_no_checking.rtf

    
    herbarium_label(dat, theme = "KFBG", outfile = "herbarium_labels_KFBG_1.1.rtf", font_size = 1.1)

#> Warning: "FAMILY" not provided for row:  8

#> [1] "17 herbarium specimen labels to create:"
#> [1] "Making label for row:  1"
#> [1] "Making label for row:  5"
#> [1] "Making label for row:  10"
#> [1] "Making label for row:  15"
#> Herbarium Labels have been saved to:
#> C:/Jinlong/github packages/herblabel/docs/reference/herbarium_labels_KFBG_1.1.rtf

    herbarium_label(dat, theme = "PE",   outfile = "herbarium_labels_PE_1.2.rtf"  , font_size = 1.2)

#> Warning: "FAMILY" not provided for row:  8

#> [1] "17 herbarium specimen labels to create:"
#> [1] "Making label for row:  1"
#> [1] "Making label for row:  5"
#> [1] "Making label for row:  10"
#> [1] "Making label for row:  15"
#> Herbarium Labels have been saved to:
#> C:/Jinlong/github packages/herblabel/docs/reference/herbarium_labels_PE_1.2.rtf

    herbarium_label(dat, theme = "KUN",  outfile = "herbarium_labels_KUN_0.8.rtf" , font_size = 0.8)

#> Warning: "FAMILY" not provided for row:  8

#> [1] "17 herbarium specimen labels to create:"
#> [1] "Making label for row:  1"
#> [1] "Making label for row:  5"
#> [1] "Making label for row:  10"
#> [1] "Making label for row:  15"
#> Herbarium Labels have been saved to:
#> C:/Jinlong/github packages/herblabel/docs/reference/herbarium_labels_KUN_0.8.rtf

    herbarium_label(dat, theme = "HU",   outfile = "herbarium_labels_HU_0.8.rtf"  , font_size = 0.8)

#> Warning: "FAMILY" not provided for row:  8

#> [1] "17 herbarium specimen labels to create:"
#> [1] "Making label for row:  1"
#> [1] "Making label for row:  5"
#> [1] "Making label for row:  10"
#> [1] "Making label for row:  15"
#> Herbarium Labels have been saved to:
#> C:/Jinlong/github packages/herblabel/docs/reference/herbarium_labels_HU_0.8.rtf

    herbarium_label(dat, spellcheck = FALSE, 
                    outfile = "herbarium_labels_no_checking.rtf"              , font_size = 1.5)

#> Warning: "FAMILY" not provided for row:  8

#> [1] "17 herbarium specimen labels to create:"
#> [1] "Making label for row:  1"
#> [1] "Making label for row:  5"
#> [1] "Making label for row:  10"
#> [1] "Making label for row:  15"
#> Herbarium Labels have been saved to:
#> C:/Jinlong/github packages/herblabel/docs/reference/herbarium_labels_no_checking.rtf

    
    ##### Labels using Sans Serif Fonts
    herbarium_label(dat, theme = "KFBG", 
                    outfile = "herbarium_labels_KFBG_Arial.rtf", 
                    font = "Arial")

#> Warning: "FAMILY" not provided for row:  8

#> [1] "17 herbarium specimen labels to create:"
#> [1] "Making label for row:  1"
#> [1] "Making label for row:  5"
#> [1] "Making label for row:  10"
#> [1] "Making label for row:  15"
#> Herbarium Labels have been saved to:
#> C:/Jinlong/github packages/herblabel/docs/reference/herbarium_labels_KFBG_Arial.rtf

    herbarium_label(dat, theme = "PE",   
                    outfile = "herbarium_labels_PE_Arial.rtf", 
                    font = "Arial")

#> Warning: "FAMILY" not provided for row:  8

#> [1] "17 herbarium specimen labels to create:"
#> [1] "Making label for row:  1"
#> [1] "Making label for row:  5"
#> [1] "Making label for row:  10"
#> [1] "Making label for row:  15"
#> Herbarium Labels have been saved to:
#> C:/Jinlong/github packages/herblabel/docs/reference/herbarium_labels_PE_Arial.rtf

    herbarium_label(dat, theme = "KUN",  
                    outfile = "herbarium_labels_KUN_Arial.rtf", 
                    font = "Arial")

#> Warning: "FAMILY" not provided for row:  8

#> [1] "17 herbarium specimen labels to create:"
#> [1] "Making label for row:  1"
#> [1] "Making label for row:  5"
#> [1] "Making label for row:  10"
#> [1] "Making label for row:  15"
#> Herbarium Labels have been saved to:
#> C:/Jinlong/github packages/herblabel/docs/reference/herbarium_labels_KUN_Arial.rtf

    herbarium_label(dat, theme = "HU",  
                    outfile = "herbarium_labels_HU_Arial.rtf", 
                    font = "Arial")

#> Warning: "FAMILY" not provided for row:  8

#> [1] "17 herbarium specimen labels to create:"
#> [1] "Making label for row:  1"
#> [1] "Making label for row:  5"
#> [1] "Making label for row:  10"
#> [1] "Making label for row:  15"
#> Herbarium Labels have been saved to:
#> C:/Jinlong/github packages/herblabel/docs/reference/herbarium_labels_HU_Arial.rtf

    herbarium_label(dat, spellcheck = FALSE, 
                    outfile = "herbarium_labels_no_checking_Arial.rtf", 
                    font = "Arial")

#> Warning: "FAMILY" not provided for row:  8

#> [1] "17 herbarium specimen labels to create:"
#> [1] "Making label for row:  1"
#> [1] "Making label for row:  5"
#> [1] "Making label for row:  10"
#> [1] "Making label for row:  15"
#> Herbarium Labels have been saved to:
#> C:/Jinlong/github packages/herblabel/docs/reference/herbarium_labels_no_checking_Arial.rtf