Parsing scientific names for a species

Parsing scientif names to FAMILY,GENUS, SPECIES, AUTHOR_OF_SPECIES, INFRASPECIFIC_RANK, INFRASPECIFIC_EPITHET, AUTHOR_OF_INFRASPECIFIC_RANK

parse_taxa(taxa)

Arguments

taxa
A string representing scientific name

Details

This functions finds the gaps in the scientific name, and parse the name to GENUS, SPECIES, AUTHOR_OF_SPECIES, INFRASPECIFIC_RANK, INFRASPECIFIC_EPITHET, AUTHOR_OF_INFRASPECIFIC_RANK. There may be errors in parsing scientific name with "f.".

Value

A dataframe containing the following columns TAXON_PARSED, GENUS_PARSED, SPECIES_PARSED, AUTHOR_OF_SPECIES_PARSED, INFRASPECIFIC_RANK_PARSED, INFRASPECIFIC_EPITHET_PARSED, AUTHOR_OF_INFRASPECIFIC_RANK_PARSED

Author

Jinlong Zhang jinlongzhang01@gmail.com

See also

Examples

parse_taxa("Acer")

#>   TAXON_PARSED GENUS_PARSED SPECIES_PARSED AUTHOR_OF_SPECIES_PARSED
#> 1         Acer         Acer                                        
#>   INFRASPECIFIC_RANK_PARSED INFRASPECIFIC_EPITHET_PARSED
#> 1                                                       
#>   AUTHOR_OF_INFRASPECIFIC_RANK_PARSED
#> 1                                    

parse_taxa("Acer caudatifolium Hayata")

#>                TAXON_PARSED GENUS_PARSED SPECIES_PARSED
#> 1 Acer caudatifolium Hayata         Acer  caudatifolium
#>   AUTHOR_OF_SPECIES_PARSED INFRASPECIFIC_RANK_PARSED
#> 1                   Hayata                          
#>   INFRASPECIFIC_EPITHET_PARSED AUTHOR_OF_INFRASPECIFIC_RANK_PARSED
#> 1                                                                 

parse_taxa("Acer caudatifolium")

#>         TAXON_PARSED GENUS_PARSED SPECIES_PARSED AUTHOR_OF_SPECIES_PARSED
#> 1 Acer caudatifolium         Acer  caudatifolium                         
#>   INFRASPECIFIC_RANK_PARSED INFRASPECIFIC_EPITHET_PARSED
#> 1                                                       
#>   AUTHOR_OF_INFRASPECIFIC_RANK_PARSED
#> 1                                    

parse_taxa("Acer sp.")

#>   TAXON_PARSED GENUS_PARSED SPECIES_PARSED AUTHOR_OF_SPECIES_PARSED
#> 1     Acer sp.         Acer            sp.                         
#>   INFRASPECIFIC_RANK_PARSED INFRASPECIFIC_EPITHET_PARSED
#> 1                                                       
#>   AUTHOR_OF_INFRASPECIFIC_RANK_PARSED
#> 1                                    

parse_taxa("Acer caudatum Wall. var. multiserratum (Maxim.) Rehd.")

#>                                            TAXON_PARSED GENUS_PARSED
#> 1 Acer caudatum Wall. var. multiserratum (Maxim.) Rehd.         Acer
#>   SPECIES_PARSED AUTHOR_OF_SPECIES_PARSED INFRASPECIFIC_RANK_PARSED
#> 1       caudatum                    Wall.                      var.
#>   INFRASPECIFIC_EPITHET_PARSED AUTHOR_OF_INFRASPECIFIC_RANK_PARSED
#> 1                multiserratum                      (Maxim.) Rehd.

parse_taxa("Acer caudatum var. multiserratum (Maxim.) Rehd.")

#>                                      TAXON_PARSED GENUS_PARSED SPECIES_PARSED
#> 1 Acer caudatum var. multiserratum (Maxim.) Rehd.         Acer       caudatum
#>   AUTHOR_OF_SPECIES_PARSED INFRASPECIFIC_RANK_PARSED
#> 1                                               var.
#>   INFRASPECIFIC_EPITHET_PARSED AUTHOR_OF_INFRASPECIFIC_RANK_PARSED
#> 1                multiserratum                      (Maxim.) Rehd.

parse_taxa("Acer caudatum Wall. var. multiserratum")

#>                             TAXON_PARSED GENUS_PARSED SPECIES_PARSED
#> 1 Acer caudatum Wall. var. multiserratum         Acer       caudatum
#>   AUTHOR_OF_SPECIES_PARSED INFRASPECIFIC_RANK_PARSED
#> 1                    Wall.                      var.
#>   INFRASPECIFIC_EPITHET_PARSED AUTHOR_OF_INFRASPECIFIC_RANK_PARSED
#> 1                multiserratum                                    

parse_taxa("Acer caudatum var. multiserratum")

#>                       TAXON_PARSED GENUS_PARSED SPECIES_PARSED
#> 1 Acer caudatum var. multiserratum         Acer       caudatum
#>   AUTHOR_OF_SPECIES_PARSED INFRASPECIFIC_RANK_PARSED
#> 1                                               var.
#>   INFRASPECIFIC_EPITHET_PARSED AUTHOR_OF_INFRASPECIFIC_RANK_PARSED
#> 1                multiserratum                                    

parse_taxa("Acer caesium Wall. ex Brandis subsp. giraldii (Pax) E. Murr.")

#>                                                   TAXON_PARSED GENUS_PARSED
#> 1 Acer caesium Wall. ex Brandis subsp. giraldii (Pax) E. Murr.         Acer
#>   SPECIES_PARSED AUTHOR_OF_SPECIES_PARSED INFRASPECIFIC_RANK_PARSED
#> 1        caesium         Wall. ex Brandis                    subsp.
#>   INFRASPECIFIC_EPITHET_PARSED AUTHOR_OF_INFRASPECIFIC_RANK_PARSED
#> 1                     giraldii                      (Pax) E. Murr.

parse_taxa("Acer caesium subsp. giraldii (Pax) E. Murr.")

#>                                  TAXON_PARSED GENUS_PARSED SPECIES_PARSED
#> 1 Acer caesium subsp. giraldii (Pax) E. Murr.         Acer        caesium
#>   AUTHOR_OF_SPECIES_PARSED INFRASPECIFIC_RANK_PARSED
#> 1                                             subsp.
#>   INFRASPECIFIC_EPITHET_PARSED AUTHOR_OF_INFRASPECIFIC_RANK_PARSED
#> 1                     giraldii                      (Pax) E. Murr.

parse_taxa("Acer caesium Wall. ex Brandis subsp. giraldii")

#>                                    TAXON_PARSED GENUS_PARSED SPECIES_PARSED
#> 1 Acer caesium Wall. ex Brandis subsp. giraldii         Acer        caesium
#>   AUTHOR_OF_SPECIES_PARSED INFRASPECIFIC_RANK_PARSED
#> 1         Wall. ex Brandis                    subsp.
#>   INFRASPECIFIC_EPITHET_PARSED AUTHOR_OF_INFRASPECIFIC_RANK_PARSED
#> 1                     giraldii                                    

parse_taxa("Acer caesium subsp. giraldii")                               

#>                   TAXON_PARSED GENUS_PARSED SPECIES_PARSED
#> 1 Acer caesium subsp. giraldii         Acer        caesium
#>   AUTHOR_OF_SPECIES_PARSED INFRASPECIFIC_RANK_PARSED
#> 1                                             subsp.
#>   INFRASPECIFIC_EPITHET_PARSED AUTHOR_OF_INFRASPECIFIC_RANK_PARSED
#> 1                     giraldii                                    

############# Not a real species #############

tax <- "   Acanthodium   furcatum ( L. f. ) Nees        "
parse_taxa(tax)

#>                          TAXON_PARSED GENUS_PARSED SPECIES_PARSED
#> 1 Acanthodium furcatum ( L. f. ) Nees  Acanthodium       furcatum
#>   AUTHOR_OF_SPECIES_PARSED INFRASPECIFIC_RANK_PARSED
#> 1           ( L. f. ) Nees                          
#>   INFRASPECIFIC_EPITHET_PARSED AUTHOR_OF_INFRASPECIFIC_RANK_PARSED
#> 1                                                                 

taxon <- "   Acanthodium   furcatum ( L. f. ) Nees        "
parse_taxa(taxon)

#>                          TAXON_PARSED GENUS_PARSED SPECIES_PARSED
#> 1 Acanthodium furcatum ( L. f. ) Nees  Acanthodium       furcatum
#>   AUTHOR_OF_SPECIES_PARSED INFRASPECIFIC_RANK_PARSED
#> 1           ( L. f. ) Nees                          
#>   INFRASPECIFIC_EPITHET_PARSED AUTHOR_OF_INFRASPECIFIC_RANK_PARSED
#> 1                                                                 

tax <- "Acanthus ebracteatus var. xiamenensis (R.T.Zhang) C.Y.Wu & C.C.Hu            "
parse_taxa(tax)

#>                                                        TAXON_PARSED
#> 1 Acanthus ebracteatus var. xiamenensis (R.T.Zhang) C.Y.Wu & C.C.Hu
#>   GENUS_PARSED SPECIES_PARSED AUTHOR_OF_SPECIES_PARSED
#> 1     Acanthus    ebracteatus                         
#>   INFRASPECIFIC_RANK_PARSED INFRASPECIFIC_EPITHET_PARSED
#> 1                      var.                  xiamenensis
#>   AUTHOR_OF_INFRASPECIFIC_RANK_PARSED
#> 1         (R.T.Zhang) C.Y.Wu & C.C.Hu

tax <- "Acanthus ebracteatus subsp. xiamenensis (R.T.Zhang) C.Y.Wu & C.C.Hu          "
parse_taxa(tax)

#>                                                          TAXON_PARSED
#> 1 Acanthus ebracteatus subsp. xiamenensis (R.T.Zhang) C.Y.Wu & C.C.Hu
#>   GENUS_PARSED SPECIES_PARSED AUTHOR_OF_SPECIES_PARSED
#> 1     Acanthus    ebracteatus                         
#>   INFRASPECIFIC_RANK_PARSED INFRASPECIFIC_EPITHET_PARSED
#> 1                    subsp.                  xiamenensis
#>   AUTHOR_OF_INFRASPECIFIC_RANK_PARSED
#> 1         (R.T.Zhang) C.Y.Wu & C.C.Hu

tax <- "Acanthus ebracteatus XXXX var. xiamenensis (R.T.Zhang) C.Y.Wu & C.C.Hu       "
parse_taxa(tax)

#>                                                             TAXON_PARSED
#> 1 Acanthus ebracteatus XXXX var. xiamenensis (R.T.Zhang) C.Y.Wu & C.C.Hu
#>   GENUS_PARSED SPECIES_PARSED AUTHOR_OF_SPECIES_PARSED
#> 1     Acanthus    ebracteatus                     XXXX
#>   INFRASPECIFIC_RANK_PARSED INFRASPECIFIC_EPITHET_PARSED
#> 1                      var.                  xiamenensis
#>   AUTHOR_OF_INFRASPECIFIC_RANK_PARSED
#> 1         (R.T.Zhang) C.Y.Wu & C.C.Hu

tax <- "Acanthus ebracteatus XXXX f. subsp. xiamenensis (R.T.Zhang) C.Y.Wu & C.C.Hu  "
parse_taxa(tax)

#>                                                                  TAXON_PARSED
#> 1 Acanthus ebracteatus XXXX f. subsp. xiamenensis (R.T.Zhang) C.Y.Wu & C.C.Hu
#>   GENUS_PARSED SPECIES_PARSED AUTHOR_OF_SPECIES_PARSED
#> 1     Acanthus    ebracteatus                  XXXX f.
#>   INFRASPECIFIC_RANK_PARSED INFRASPECIFIC_EPITHET_PARSED
#> 1                    subsp.                  xiamenensis
#>   AUTHOR_OF_INFRASPECIFIC_RANK_PARSED
#> 1         (R.T.Zhang) C.Y.Wu & C.C.Hu

tax <- "Acanthus ebracteatus XXXX  subsp. xiamenensis (R.T.Zhang) C.Y.Wu & C.C.Hu f. "
parse_taxa(tax)

#>                                                                  TAXON_PARSED
#> 1 Acanthus ebracteatus XXXX subsp. xiamenensis (R.T.Zhang) C.Y.Wu & C.C.Hu f.
#>   GENUS_PARSED SPECIES_PARSED AUTHOR_OF_SPECIES_PARSED
#> 1     Acanthus    ebracteatus                     XXXX
#>   INFRASPECIFIC_RANK_PARSED INFRASPECIFIC_EPITHET_PARSED
#> 1                    subsp.                  xiamenensis
#>   AUTHOR_OF_INFRASPECIFIC_RANK_PARSED
#> 1      (R.T.Zhang) C.Y.Wu & C.C.Hu f.

tax <- "Acanthus ebracteatus XXXX f. subsp. xiamenensis (R.T.Zhang) C.Y.Wu & C.C.Hu  "
parse_taxa(tax)

#>                                                                  TAXON_PARSED
#> 1 Acanthus ebracteatus XXXX f. subsp. xiamenensis (R.T.Zhang) C.Y.Wu & C.C.Hu
#>   GENUS_PARSED SPECIES_PARSED AUTHOR_OF_SPECIES_PARSED
#> 1     Acanthus    ebracteatus                  XXXX f.
#>   INFRASPECIFIC_RANK_PARSED INFRASPECIFIC_EPITHET_PARSED
#> 1                    subsp.                  xiamenensis
#>   AUTHOR_OF_INFRASPECIFIC_RANK_PARSED
#> 1         (R.T.Zhang) C.Y.Wu & C.C.Hu

tax <- "Acanthodium furcatum f. variagata Nees      "
parse_taxa(tax)

#>                             TAXON_PARSED GENUS_PARSED SPECIES_PARSED
#> 1 Acanthodium furcatum f. variagata Nees  Acanthodium       furcatum
#>   AUTHOR_OF_SPECIES_PARSED INFRASPECIFIC_RANK_PARSED
#> 1                                                 f.
#>   INFRASPECIFIC_EPITHET_PARSED AUTHOR_OF_INFRASPECIFIC_RANK_PARSED
#> 1                    variagata                                Nees

tax <- "Acanthodium furcatum (L. f. ) Nees          "
parse_taxa(tax)

#>                         TAXON_PARSED GENUS_PARSED SPECIES_PARSED
#> 1 Acanthodium furcatum (L. f. ) Nees  Acanthodium       furcatum
#>   AUTHOR_OF_SPECIES_PARSED INFRASPECIFIC_RANK_PARSED
#> 1            (L. f. ) Nees                          
#>   INFRASPECIFIC_EPITHET_PARSED AUTHOR_OF_INFRASPECIFIC_RANK_PARSED
#> 1                                                                 

tax <- "Acanthus ebracteatus subsp. ebracteatus"
parse_taxa(tax)

#>                              TAXON_PARSED GENUS_PARSED SPECIES_PARSED
#> 1 Acanthus ebracteatus subsp. ebracteatus     Acanthus    ebracteatus
#>   AUTHOR_OF_SPECIES_PARSED INFRASPECIFIC_RANK_PARSED
#> 1                                             subsp.
#>   INFRASPECIFIC_EPITHET_PARSED AUTHOR_OF_INFRASPECIFIC_RANK_PARSED
#> 1                  ebracteatus